bots/athena/requirements.txt
python-telegram-bot
google-generativeai
python-dotenv
supabase
uvicorn
fastapi
requests
